async requests, pagination, storing file

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 const mongoose = require('mongoose');
+const fileHelper = require('../util/file');
 
 const { validationResult } = require('express-validator/check');
 
@@ -17,9 +18,26 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
+    console.log(title);
+    console.log(image);
+    if(!image) {
+        return res.status(422).render('admin/edit-product', { 
+            pageTitle: 'Add Product', 
+            path: '/admin/add-product',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description
+            },
+            errorMessage: 'Attached file is not an image',
+            validationErrors: []
+        }); 
+    }
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
         console.log(errors.array());
@@ -39,15 +57,20 @@ exports.postAddProduct = (req, res, next) => {
         }); 
     }
     console.log(errors.array());
+
+    let imageUrl = image.path;
     
+    console.log(imageUrl);
+
     const product = new Product({
         //_id: new mongoose.Types.ObjectId('5d0a0c991ccc3b1938cf0fad'),
         title: title,
         price: price,
-        description: description,
         imageUrl: imageUrl,
+        description: description,
         userId: req.user
     });
+    
     product
         .save()
         .then(result => {
@@ -109,7 +132,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
@@ -120,7 +143,6 @@ exports.postEditProduct = (req, res, next) => {
           hasError: true,
           product: {
               title: updatedTitle,
-              imageUrl: updatedImageUrl,
               price: updatedPrice,
               description: updatedDesc,
               _id: prodId
@@ -137,7 +159,10 @@ exports.postEditProduct = (req, res, next) => {
         product.title = updatedTitle;
         product.price = updatedPrice;
         product.description = updatedDesc;
-        product.imageUrl = updatedImageUrl;
+        if(image) {
+          fileHelper.deleteFile(product.imageUrl);
+          product.imageUrl = image.path;  
+        }
         return product.save().then(result => {
             console.log('UPDATED PRODUCT!');
             res.redirect('/admin/products');
@@ -171,16 +196,21 @@ exports.getProducts = (req, res, next) => {
       });
 }
 
-exports.postDeleteProduct = (req, res, next) => {
-    const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id})
-        .then(() => {
-            console.log('DESTROYED PRODUCT');
-            res.redirect('/admin/products');
-        })
-        .catch(err => {
-            const error = new Error(err);
-            error.httpStatusCode = 500;
-            return next(error);
-          });
+exports.deleteProduct = (req, res, next) => {
+    const prodId = req.params.productId;
+
+    Product.findById(prodId).then(product => {
+        if(!product) {
+            return next(new Error('Product not found.'));
+        }
+        fileHelper.deleteFile(product.imageUrl);
+        return Product.deleteOne({ _id: prodId, userId: req.user._id});
+    })
+    .then(() => {
+        console.log('DESTROYED PRODUCT');
+        res.status(200).json({ message: 'Success!'});
+    })
+    .catch(err => {
+        res.status(500).json({message: 'Deleting product failed.'});
+      });
 }
